@@ -1,13 +1,10 @@
 name: ms_map_utils
 description: A dart package that implement a lot of util functions for a Map
-version: 0.7.0
+version: 1.0.0
 homepage: https://github.com/Marcus-Software/ms_map_utils
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
-
-dependencies:
-  meta: '>=1.0.0 <2.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dev_dependencies:
-  test: ^1.9.2
+  test: ^1.16.7
