@@ -7,9 +7,9 @@ part of '../ms_map_utils.dart';
 Map<K, V> compact<K, V>(Map map, [bool newMap = false]) =>
     (newMap ? Map<K, V>.from(map) : map).cast<K, V>()
       ..removeWhere((_, value) {
-        if (value is Map && value != null) {
+        if (value is Map) {
           compact(value, newMap);
-        } else if (value is List<Map> && value != null) {
+        } else if (value is List<Map>) {
           value.forEach((mapInList) {
             compact(mapInList, newMap);
           });
