@@ -30,10 +30,5 @@ void main() {
     expect(newThing, null);
   });
 
-  test('must return throw assertion error if doWork is null', () {
-    final testMap = {'key1': 'value1', 'key2': 'value2'};
-    // ignore: missing_required_param
-    expect(() => doIfContains<List<String>>(testMap, 'key3'),
-        throwsA(TypeMatcher<AssertionError>()));
-  });
+
 }
